feat(core): add new useObserveDocument hook
