+++ packages/sanity/src/core/preview/index.ts
@@ -3,6 +3,7 @@ export * from './components/PreviewLoader'
 export * from './components/SanityDefaultPreview'
 export * from './documentPreviewStore'
 export * from './types'
+export * from './useObserveDocument'
 export * from './useValuePreview'
 export {getPreviewPaths} from './utils/getPreviewPaths'
 export {getPreviewStateObservable} from './utils/getPreviewStateObservable'
